add comprehantion

+++ gain_train_dataset.py
@@ -41,17 +41,19 @@ def normalize_8_bit(img):
 
 
 if __name__ == "__main__":
-    tiff_reconstructed_img_dir = r"C:\Users\ankha\Desktop\geckons"
-    N = 100
+    tiff_reconstructed_img_dir = r"C:\Users\ankha\Desktop\diplom\geckons"
+    already_placed = os.listdir(IMGS_FOLDER)
+    N = 3
 
     img_tiff_names = choose_random_N_tiff_images(tiff_reconstructed_img_dir, N)
     
     for img_tiff_name in tqdm(img_tiff_names):
-        img_tiff_path_orig = os.path.join(tiff_reconstructed_img_dir, img_tiff_name)
+        if img_tiff_name not in already_placed:
+            img_tiff_path_orig = os.path.join(tiff_reconstructed_img_dir, img_tiff_name)
 
-        img = normalize_8_bit(cv2.imread(img_tiff_path_orig, -1))
-        img_png_path = os.path.join(IMGS_PNG_FOLDER, "".join(img_tiff_name.split(".")[:-1])+"."+"png")
-        img_tiff_path = os.path.join(IMGS_FOLDER, img_tiff_name)
+            img = normalize_8_bit(cv2.imread(img_tiff_path_orig, -1))
+            img_png_path = os.path.join(IMGS_PNG_FOLDER, "".join(img_tiff_name.split(".")[:-1])+"."+"png")
+            img_tiff_path = os.path.join(IMGS_FOLDER, img_tiff_name)
 
-        cv2.imwrite(img_png_path, img)
-        cv2.imwrite(img_tiff_path, img)
+            cv2.imwrite(img_png_path, img)
+            cv2.imwrite(img_tiff_path, img)
